chabgeEntitySharedServiceを修正 #27
