+++ src/main/resources/application.properties
@@ -3,3 +3,4 @@ spring.datasource.url=jdbc:mysql://${MYSQL_HOST:localhost}:3306/boyaki_dining
 spring.datasource.username=boyakiUser
 spring.datasource.password=nyousanti
 spring.datasource.driver-class-name =com.mysql.cj.jdbc.Driver
+spring.jpa.open-in-view= true
